fix: simple css stack
I wanted to modify the css and as I am not used to sass, less, etc.
nor did I found any help in the readme, I decided to do this PR.

It greatly simplifies the CSS stack to only one readable pure css
file.

Style is unchanged.

I feel we wanted to provide "tooling" to the instant search users but
I do feel it was too much and complexifies the stack.

What we want is show to them how to build an instant search, not how
to inherit colors from a bootstrap sass file.

+ load bootstrap from cdn (js & css)
+ use revamped bootstrap-slider
+ use latest jquery